Prettier format

+++ pages/api/events/[id].js
@@ -4,8 +4,6 @@ import {
   deleteEventById,
 } from '@/lib/prisma/events';
 const handler = async (req, res) => {
-  if (req.query.secret !== process.env.SECRET_TOKEN) return res.status(401).json({message:"Invalid Token"})
-
   const { id } = req.query;
   if (req.method === 'GET') {
     try {
@@ -16,6 +14,9 @@ const handler = async (req, res) => {
       return res.status(400).json({ error: error.message });
     }
   }
+  if (req.query.secret !== process.env.SECRET_TOKEN)
+    return res.status(401).json({ message: 'Invalid Token' });
+
   if (req.method === 'DELETE') {
     deleteEventById(id);
     await res.revalidate('/');
