+++ pages/api/posts/[id].js
@@ -1,7 +1,5 @@
 import { getPostById, updatePostById, deletePostById } from 'lib/prisma/news';
 const handler = async (req, res) => {
-  if (req.query.secret !== process.env.SECRET_TOKEN) return res.status(401).json({message:"Invalid Token"})
-
   const { id } = req.query;
   if (req.method === 'GET') {
     try {
@@ -12,6 +10,9 @@ const handler = async (req, res) => {
       return res.status(400).json({ error: error.message });
     }
   }
+  if (req.query.secret !== process.env.SECRET_TOKEN)
+    return res.status(401).json({ message: 'Invalid Token' });
+
   if (req.method === 'DELETE') {
     deletePostById(id);
     await res.revalidate('/');
